+++ rest-client/src/main/java/restwars/restapi/EventResource.java
@@ -10,5 +10,10 @@ import javax.ws.rs.core.MediaType;
 @Produces(MediaType.APPLICATION_JSON)
 public interface EventResource {
     @GET
+    EventsResponse getEvents();
+    @GET
     EventsResponse getEvents(@QueryParam("since") long round);
+
+    @GET
+    EventsResponse getEvents(@QueryParam("since") long round, @QueryParam("max") int max);
 }