+++ service/src/main/java/restwars/service/event/EventService.java
@@ -17,6 +17,16 @@ public interface EventService {
     List<EventWithPlanet> findSince(UUID playerId, long round);
 
     /**
+     * Finds all events for a player which have occurred since the given round. Returns a given maximum number of events.
+     *
+     * @param playerId Id of the player.
+     * @param round    Round (inclusive).
+     * @param max      Maximum number of events.
+     * @return All events which have occurred since the given round.
+     */
+    List<EventWithPlanet> findSinceMax(UUID playerId, long round, int max);
+
+    /**
      * Creates a research completed event.
      *
      * @param playerId Id of the player.